pyproject.toml
[tool.poetry]
name = "etl-app"
version = "0.1.0"
description = ""
authors = ["Travis <16505078+TravJav@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
zipfile2 = "^0.0.12"
pandas = "^2.2.2"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
